requirements.txt
Flask-Mail~=0.10.0
pyodbc~=5.1.0
Jinja2~=3.1.4
Flask~=3.0.3
python-dotenv~=1.0.1